Updated API Removed linting error
