+++ src/CustomVisuals/Published/HorizontalFunnel/package.json
@@ -3,15 +3,17 @@
   "devDependencies": {
     "@types/d3": "^3.5.36",
     "@types/jquery": "^3.3.29",
-    "tslint": "^5.16.0",
-    "tslint-microsoft-contrib": "^6.1.1",
+    "@types/node": "10.3.2",
+    "tslint": "^5.20.1",
+    "tslint-microsoft-contrib": "^6.2.0",
     "typescript": "3.4.5"
   },
   "dependencies": {
     "d3": "^3.5.5",
     "jquery": "^3.4.0",
-    "lodash": "4.17.11",
+    "lodash": "4.17.15",
     "powerbi-models": "1.1.0",
+    "powerbi-visuals-tools": "^3.1.7",
     "powerbi-visuals-utils-chartutils": "1.9.0",
     "powerbi-visuals-utils-colorutils": "1.2.0",
     "powerbi-visuals-utils-dataviewutils": "1.5.0",
@@ -24,6 +26,7 @@
   "scripts": {
     "package": "pbiviz package",
     "start": "pbiviz start",
-    "tsc": "tsc"
+    "tsc": "tsc",
+    "lint": "tslint -c tslint.json -p tsconfig.json"
   }
 }
