+++ src/CustomVisuals/Published/HorizontalFunnel/tsconfig.json
@@ -10,7 +10,7 @@
   "files": [
     "node_modules/powerbi-models/dist/models-noexports.d.ts",
     ".api/v2.5.0/PowerBI-visuals.d.ts",
-    "src/visual.ts",
+    "src/HorizontalFunnel.ts",
     "node_modules/powerbi-visuals-utils-dataviewutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-formattingutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-interactivityutils/lib/index.d.ts",
