+++ src/CustomVisuals/Published/HorizontalFunnel/tslint.json
@@ -1,11 +1,10 @@
 {
-    "extends": "./node_modules/tslint/lib/configs/recommended",
+    "extends": "tslint-microsoft-contrib/recommended",
     "rulesDirectory": [
     "node_modules/tslint-microsoft-contrib"
     ],
     "rules": {
     "no-internal-module": false,
-    "no-namespace": false,
-    "trailing-comma": [true, {"multiline": "ignore", "singleline": "never"}]
+    "no-namespace": false
     }
    }
